flavour changes

+++ app/Http/Controllers/FlavourController.php
@@ -13,10 +13,10 @@ class FlavourController extends Controller
 
     public function __construct()
     {
-         $this->user = Auth::user();   
-         
+        $this->user = Auth::user();
     }
-    public function index(){
+    public function index()
+    {
         // return $this->user;   
         return view('admin.flavour.index');
     }
@@ -26,21 +26,21 @@ class FlavourController extends Controller
         $start = $request->get('start');
         $length = $request->get('length');
         $search = $request->search['value'];
-        $flavour_count = Flavour::when($search , function($q) use( $search){
-            $q->where('flavour_title' , 'LIKE' , "%$search%");
-        })->where('is_hidden' , '0')->count();
-        $flavour = Flavour::when($search , function($q) use( $search){
-            $q->where('flavour_title' , 'LIKE' , "%$search%");
+        $flavour_count = Flavour::when($search, function ($q) use ($search) {
+            $q->where('flavour_title', 'LIKE', "%$search%");
+        })->where('is_hidden', '0')->count();
+        $flavour = Flavour::when($search, function ($q) use ($search) {
+            $q->where('flavour_title', 'LIKE', "%$search%");
         });
 
-        $flavour = $flavour->where('is_hidden' , '0')->skip((int)$start)->take((int)$length)->get();  
+        $flavour = $flavour->where('is_hidden', '0')->skip((int)$start)->take((int)$length)->get();
         $data = array(
             'draw' => $draw,
             'recordsTotal' => $flavour_count,
             'recordsFiltered' => $flavour_count,
             'data' => $flavour,
         );
-       return response()->json($data);
+        return response()->json($data);
     }
     public function create(Request $request)
     {
@@ -51,21 +51,25 @@ class FlavourController extends Controller
         $flavour = new  Flavour();
         $flavour->flavour_title = $request->title;
         $flavour->flavour_description = $request->description;
-       
+
         $flavour->save();
         return redirect('/flavour/index');
     }
-    public function delete(Request $request , $id)
-    {   
-        $flavour =Flavour::find(base64_decode($id));
-        $flavour->delete();
-        return redirect('/flavour/index')->with('msg' , 'flavour Deleted Successfully');
-    }
-    public function edit(Request $request , $id)
-    {  
+    public function delete(Request $request, $id)
+    {
         $flavour = Flavour::find(base64_decode($id));
-      
-        return view('admin.flavour.edit' ,[
+        if ($flavour) {
+
+            $flavour->delete();
+        }
+        // return $flavour;
+        return redirect('/flavour/index')->with('msg', 'flavour Deleted Successfully');
+    }
+    public function edit(Request $request, $id)
+    {
+        $flavour = Flavour::find(base64_decode($id));
+
+        return view('admin.flavour.edit', [
             'flavour' =>  $flavour,
         ]);
     }
@@ -74,7 +78,7 @@ class FlavourController extends Controller
         $flavour = Flavour::find($request->id);
         $flavour->flavour_title = $request->title;
         $flavour->flavour_description = $request->description;
-        
+
         $flavour->save();
         return redirect('/flavour/index');
     }
